feat: add support to custom entry
Add support to a custom deploy entry. The developer will specify
the source, the methods, and if the source or the attribute that
is obtained from the source is shared or not. If it is shared,
a singleton will be used.

The _custom entry will support unimplemented resolution methods
in the GraphQL VSS Server Libs as well as resollution methods
that use proprietary code like COD.

Proposed custom layer:

NameOfTheSpec:
  _custom:
    origin: InfoProvider
    sharedOrigin: true
    sharedAttribute: false
    methods:
       read:
          source:
            attribute: InfoWeWant
           conversionFunction: fromAttributeAtOriginToNameOfTheSpec

* _custom: Key to generate a “generic” resolver;
* origin: The tool or protocol that is used to get the information,
ex Cod, Http, etc
* sharedOrigin: True if it is expected that multiple resolver
functions will get information from the same origin; a singleton is
used to “load” the process that will call the information, similarly
to what we do with CommonAPI proxies;
* sharedAttribute: True if it is expected that multiple resolver
functions will get information from the same attribute; a singleton
to cache its data during a resolution will be used.
* methods: Work similarly to what we have in the _francaIDL entries.

Change-Id: Iffaeec005321c1f222edf88b9f87343a14a954bd